Spider for creating input and output files for competitions created
Using Python3 , BeautifulSoup4 , selenium and chromium webdriver, created a
spider which will go through the pages of competition dashboard and extract the
input and output test cases and write it to local hard drive for use. This saves a lot of time during competitions.

PS : Now, only Google codejam support is present.